pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ragcore"
version = "1.0.0"
description = "RAG-as-a-service — MCP + OpenAI-compatible API for any AI agent"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "httpx>=0.27.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "fastmcp>=2.0.0",
    "loguru>=0.7.0",
    "pypdf>=4.0.0",
    "pandas>=2.0.0",
    "python-docx>=1.1.0",
    "python-multipart>=0.0.9",
    "numpy>=1.26.0",
    "rank-bm25>=0.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",
]
graph = ["networkx>=3.0", "spacy>=3.0"]
multimodal = ["pytesseract>=0.3", "Pillow>=10.0"]

[project.scripts]
ragcore = "ragcore.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["ragcore*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
